sabotage Yizar

+++ Projects/P1/src/Ghost.java
@@ -35,7 +35,7 @@ public class Ghost{
 		if(validMoves.isEmpty()) {
 			return false;
 		} else {
-			this.myLoc = validMoves.get((int)(Math.random()*validMoves.size()));
+			this.myLoc = validMoves.get((Math.random()*validMoves.size()));
 			this.shift = this.myLoc.unshift(this.myLoc);
 			this.myMap.move(this.myName, this.myLoc, Map.Type.GHOST);
 			return true;
